Refactor: clean up controller
