+++ package.json
@@ -122,6 +122,10 @@
                     "when": "resourceScheme == file && resourceExtname == .grc"
                 },
                 {
+                    "command": "gnuradio-integration.createBlock",
+                    "when": "gnuradio-integration.moduleFound"
+                },
+                {
                     "command": "gnuradio-integration.createPythonBindings",
                     "when": "gnuradio-integration.moduleFound"
                 },
