+++ src/extension.ts
@@ -1,3 +1,5 @@
+'use strict';
+
 import * as vscode from 'vscode';
 import { GNURadioController } from './controller';
